6_PROJECT_next_word_LSTM/requirements.txt
tensorflow
pandas 
numpy 
scikit-learn
tensorboard
matplotlib
streamlit
scikeras
tensorflow_datasets 